ESQL: Move `rule` package out of QL clone (#113637)
We're trying to flatten away our clone of QL called `esql-core` so that
it's easier to work on ESQL directly. This bit moves cleanly.
